Add DBSIZE
While adding the command I noticed several problems with inconsistent
deletion. To fix it and prevent it from happening again, when a command
removes an element from a collection, if the collection is now empty,
the key is completely removed. This is consistent with redis behavior.

Notice that all functions that remove elements (including some that
mutate existing elements, but actually remove one and add another) now
must receive a bunch of additional parameters to fulfil the potencial
deletion of the key.