Calculate btree size
Instead of it being provided as an argument, it is now calculated
automatically based on the type of btree.

This raised a problem that when the number of elements in a node was odd,
the btree would have off-by-one problems. Current solution is to
approximate to the immediately lower even number. Alternatively,
btrees could work for odd number of elements.