Fix valgrind error on tests
Some tests use a really small page size. Adding a database for the lua
scripts was writing after the limit. This patch will make it so page
numbers are not written when they are 0, which we can do because we
memset to 0. It's also a small performace improvement, and fixes the bug
as a side effect.